move routing file into libs
